auth_api/src/requirements.txt
redis==4.4.4
fastapi==0.100.1
orjson==3.9.5
pydantic[email]==1.9.1
uvicorn==0.23.2
python-dotenv==1.0
gunicorn==21.2.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
asyncpg==0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.11.3
typer[all]==0.9.0
aiohttp==3.8.5
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-instrumentation-redis==0.38b0
opentelemetry-instrumentation-sqlalchemy==0.38b0
opentelemetry-exporter-jaeger==1.17.0
asgi-correlation-id==4.2.0
starlette_context==0.3.6